fix warning, add hf constant, and refactor losses
